rpma: add struct rpma_cq and related functions
1) add struct rpma_cq and rpma_cq_* functions.
2) take use of struct rpma_cq and rpma_cq_* functions.
