+++ src/conn_req.c
@@ -1,5 +1,6 @@
 // SPDX-License-Identifier: BSD-3-Clause
 /* Copyright 2020-2021, Intel Corporation */
+/* Copyright 2021, Fujitsu */
 
 /*
  * conn_req.c -- librpma connection-request-related implementations
@@ -27,17 +28,14 @@ struct rpma_conn_req {
 	struct rdma_cm_event *edata;
 	/* CM ID of the connection request */
 	struct rdma_cm_id *id;
-	/* completion queue of the CM ID */
-	struct ibv_cq *cq;
+	/* rpma_cq object */
+	struct rpma_cq *cq;
 
 	/* private data of the CM ID (incoming only) */
 	struct rpma_conn_private_data data;
 
 	/* a parent RPMA peer of this request - needed for derivative objects */
 	struct rpma_peer *peer;
-
-	/* completion event channel - copy for convenience */
-	struct ibv_comp_channel *channel;
 };
 
 /*
@@ -53,40 +51,19 @@ rpma_conn_req_from_id(struct rpma_peer *peer, struct rdma_cm_id *id,
 {
 	int ret = 0;
 
-	/* create a completion channel */
-	struct ibv_comp_channel *channel = ibv_create_comp_channel(id->verbs);
-	if (channel == NULL) {
-		RPMA_LOG_ERROR_WITH_ERRNO(errno, "ibv_create_comp_channel()");
-		return RPMA_E_PROVIDER;
-	}
-
 	/* read CQ size from the configuration */
 	int cqe;
 	(void) rpma_conn_cfg_get_cqe(cfg, &cqe);
 
-	/* create a CQ */
-	struct ibv_cq *cq = ibv_create_cq(id->verbs, cqe,
-				NULL /* cq_context */,
-				channel /* channel */,
-				0 /* comp_vector */);
-	if (cq == NULL) {
-		RPMA_LOG_ERROR_WITH_ERRNO(errno, "ibv_create_cq()");
-		ret = RPMA_E_PROVIDER;
-		goto err_destroy_comp_channel;
-	}
-
-	/* request for the next completion on the completion channel */
-	errno = ibv_req_notify_cq(cq, 0 /* all completions */);
-	if (errno) {
-		RPMA_LOG_ERROR_WITH_ERRNO(errno, "ibv_req_notify_cq()");
-		ret = RPMA_E_PROVIDER;
-		goto err_destroy_cq;
-	}
+	struct rpma_cq *cq = NULL;
+	ret = rpma_cq_new(id->verbs, cqe, &cq);
+	if (ret)
+		return ret;
 
 	/* create a QP */
 	ret = rpma_peer_create_qp(peer, id, cq, cfg);
 	if (ret)
-		goto err_destroy_cq;
+		goto err_rpma_cq_delete;
 
 	*req_ptr = (struct rpma_conn_req *)malloc(sizeof(struct rpma_conn_req));
 	if (*req_ptr == NULL) {
@@ -100,18 +77,14 @@ rpma_conn_req_from_id(struct rpma_peer *peer, struct rdma_cm_id *id,
 	(*req_ptr)->data.ptr = NULL;
 	(*req_ptr)->data.len = 0;
 	(*req_ptr)->peer = peer;
-	(*req_ptr)->channel = channel;
 
 	return 0;
 
 err_destroy_qp:
 	rdma_destroy_qp(id);
 
-err_destroy_cq:
-	(void) ibv_destroy_cq(cq);
-
-err_destroy_comp_channel:
-	(void) ibv_destroy_comp_channel(channel);
+err_rpma_cq_delete:
+	(void) rpma_cq_delete(&cq);
 
 	return ret;
 }
@@ -119,7 +92,7 @@ err_destroy_comp_channel:
 /*
  * rpma_conn_req_accept -- call rdma_accept()+rdma_ack_cm_event(). If succeeds
  * request re-packing the connection request to a connection object. Otherwise,
- * rdma_disconnect()+rdma_destroy_qp()+ibv_destroy_cq() to destroy
+ * rdma_disconnect()+rdma_destroy_qp()+rpma_cq_delete() to destroy
  * the unsuccessful connection request.
  *
  * ASSUMPTIONS
@@ -160,8 +133,7 @@ err_conn_disconnect:
 
 err_conn_req_delete:
 	rdma_destroy_qp(req->id);
-	(void) ibv_destroy_cq(req->cq);
-	(void) ibv_destroy_comp_channel(req->channel);
+	(void) rpma_cq_delete(&req->cq);
 
 	return ret;
 }
@@ -169,7 +141,7 @@ err_conn_req_delete:
 /*
  * rpma_conn_req_connect_active -- call rdma_connect(). If succeeds request
  * re-packing the connection request to a connection object. Otherwise,
- * rdma_destroy_qp()+ibv_destroy_cq()+rdma_destroy_id() to destroy
+ * rdma_destroy_qp()+rpma_cq_delete()+rdma_destroy_id() to destroy
  * the unsuccessful connection request.
  *
  * ASSUMPTIONS
@@ -185,8 +157,7 @@ rpma_conn_req_connect_active(struct rpma_conn_req *req,
 	ret = rpma_conn_new(req->peer, req->id, req->cq, &conn);
 	if (ret) {
 		rdma_destroy_qp(req->id);
-		(void) ibv_destroy_cq(req->cq);
-		(void) ibv_destroy_comp_channel(req->channel);
+		(void) rpma_cq_delete(&req->cq);
 		(void) rdma_destroy_id(req->id);
 		return ret;
 	}
@@ -210,19 +181,7 @@ rpma_conn_req_connect_active(struct rpma_conn_req *req,
 static int
 rpma_conn_req_reject(struct rpma_conn_req *req)
 {
-	int ret = 0;
-
-	errno = ibv_destroy_cq(req->cq);
-	if (errno) {
-		RPMA_LOG_ERROR_WITH_ERRNO(errno, "ibv_destroy_cq()");
-		ret = RPMA_E_PROVIDER;
-	}
-
-	errno = ibv_destroy_comp_channel(req->channel);
-	if (!ret && errno) {
-		RPMA_LOG_ERROR_WITH_ERRNO(errno, "ibv_destroy_comp_channel()");
-		ret = RPMA_E_PROVIDER;
-	}
+	int ret = rpma_cq_delete(&req->cq);
 
 	if (rdma_reject(req->id,
 			NULL /* private data */,
@@ -252,27 +211,16 @@ rpma_conn_req_reject(struct rpma_conn_req *req)
 static int
 rpma_conn_req_destroy(struct rpma_conn_req *req)
 {
-	errno = ibv_destroy_cq(req->cq);
-	if (errno) {
-		RPMA_LOG_ERROR_WITH_ERRNO(errno, "ibv_destroy_cq()");
-		(void) ibv_destroy_comp_channel(req->channel);
-		(void) rdma_destroy_id(req->id);
-		return RPMA_E_PROVIDER;
-	}
-
-	errno = ibv_destroy_comp_channel(req->channel);
-	if (errno) {
-		RPMA_LOG_ERROR_WITH_ERRNO(errno, "ibv_destroy_comp_channel()");
-		(void) rdma_destroy_id(req->id);
-		return RPMA_E_PROVIDER;
-	}
+	int ret = rpma_cq_delete(&req->cq);
 
 	if (rdma_destroy_id(req->id)) {
-		RPMA_LOG_ERROR_WITH_ERRNO(errno, "rdma_destroy_id()");
-		return RPMA_E_PROVIDER;
+		if (!ret) {
+			RPMA_LOG_ERROR_WITH_ERRNO(errno, "rdma_destroy_id()");
+			ret = RPMA_E_PROVIDER;
+		}
 	}
 
-	return 0;
+	return ret;
 }
 
 /* internal librpma API */
