+++ src/peer.c
@@ -1,5 +1,6 @@
 // SPDX-License-Identifier: BSD-3-Clause
 /* Copyright 2020-2021, Intel Corporation */
+/* Copyright 2021, Fujitsu */
 
 /*
  * peer.c -- librpma peer-related implementations
@@ -87,7 +88,7 @@ rpma_peer_usage_to_access(struct rpma_peer *peer, int usage)
  */
 int
 rpma_peer_create_qp(struct rpma_peer *peer, struct rdma_cm_id *id,
-		struct ibv_cq *cq, const struct rpma_conn_cfg *cfg)
+		struct rpma_cq *cq, const struct rpma_conn_cfg *cfg)
 {
 	if (peer == NULL || id == NULL || cq == NULL)
 		return RPMA_E_INVAL;
@@ -98,10 +99,12 @@ rpma_peer_create_qp(struct rpma_peer *peer, struct rdma_cm_id *id,
 	(void) rpma_conn_cfg_get_sq_size(cfg, &sq_size);
 	(void) rpma_conn_cfg_get_rq_size(cfg, &rq_size);
 
+	struct ibv_cq *ibv_cq = rpma_cq_get_ibv_cq(cq);
+
 	struct ibv_qp_init_attr qp_init_attr;
 	qp_init_attr.qp_context = NULL;
-	qp_init_attr.send_cq = cq;
-	qp_init_attr.recv_cq = cq;
+	qp_init_attr.send_cq = ibv_cq;
+	qp_init_attr.recv_cq = ibv_cq;
 	qp_init_attr.srq = NULL;
 	qp_init_attr.cap.max_send_wr = sq_size;
 	qp_init_attr.cap.max_recv_wr = rq_size;
